Implement getProgress with a file
The idea is to use a file created automatically before executing a task to keep
track of the progress of a task. The path to this file is then accessible by
tasks through a new variable called PA_TASK_PROGRESS_FILE.

Each time the content of the file pointed by PA_TASK_PROGRESS_FILE is updated,
the event is detected by using a WatchService. If the new value is valid, then
it is kept in memory by the TaskLauncher so that any call to the getProgress()
method returns the last value which was read.

Change-Id: Ie6a7b058762591951e2fd8881a8fe30aec6da41b